% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example Data}
\format{
A data frame with X rows and Y columns (描述数据的具体格式).
}
\source{
Describe the source of the data, if applicable
}
\usage{
example_data
}
\description{
A detailed description of the example_data dataset.
}
\keyword{datasets}
