% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Check.R
\name{Format_Check}
\alias{Format_Check}
\title{Format Check for Input Data}
\usage{
Format_Check(input)
}
\arguments{
\item{input}{The input dataset, expected to be a data frame.}
}
\value{
Prints a message indicating whether the input data format is correct, if any columns are missing,
or if the column order is incorrect. Additionally, checks for NA values in the 'FC' column
and prints a message if any are found.
}
\description{
Checks if the input data contains all the required columns in the correct order and validates the format.
The function also checks for NA values in the 'FC' column.
}
