% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KESA_plot.R
\name{KESA_plot}
\alias{KESA_plot}
\title{Generate KESA Plot}
\usage{
KESA_plot(x)
}
\arguments{
\item{x}{A data frame containing the columns \code{m}, \code{mS}, \code{z.score}, \code{p.value}, \code{GENE},
and \code{sigmarker}. It assumes that \code{x} is a pre-processed data frame.}
}
\value{
A ggplot object representing the KESA plot. The plot includes a gradient fill based on
\code{z.score} values, with a vertical line at \code{x = 0} and customized theme settings.
}
\description{
This function generates a bar plot with vertical lines and gradient color fill
based on \code{z.score}. It is designed to display the distribution of \code{z.score}
across different \code{GENE} levels, highlighting significant markers.
}
